+++ partner/src/epilot/models/operations/resendpartnerinvitation.py
@@ -19,7 +19,7 @@ class ResendPartnerInvitationRequestBodyLanguage(str, Enum):
 
 @dataclasses.dataclass
 class ResendPartnerInvitationRequestBody:
-    language: Optional[ResendPartnerInvitationRequestBodyLanguage] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
+    language: Optional[ResendPartnerInvitationRequestBodyLanguage] = dataclasses.field(default=ResendPartnerInvitationRequestBodyLanguage.EN, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
     r"""Language for partner invitation email"""