+++ partner/src/epilot/models/shared/assignableorganization.py
@@ -4,9 +4,8 @@ from __future__ import annotations
 import dataclasses
 from ..shared import addressgeolocation as shared_addressgeolocation
 from dataclasses_json import Undefined, dataclass_json
-from enum import Enum
 from epilot import utils
-from typing import Optional
+from typing import Final, Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -18,9 +17,6 @@ class AssignableOrganizationImageURI:
     
 
 
-class AssignableOrganizationType(str, Enum):
-    PARTNER_ORGANIZATION = 'partner_organization'
-
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
@@ -29,7 +25,7 @@ class AssignableOrganization:
     display_name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('display_name') }})
     org_id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('org_id') }})
     partner_id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('partner_id') }})
-    type: AssignableOrganizationType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    TYPE: Final[str] = dataclasses.field(default='partner_organization', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     activity_radius: Optional[float] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('activity_radius'), 'exclude': lambda f: f is None }})
     r"""Activity radius, in km, the partner is operating in"""
     email: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('email'), 'exclude': lambda f: f is None }})