+++ partner/src/epilot/models/shared/assignableuser.py
@@ -3,9 +3,8 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
-from enum import Enum
 from epilot import utils
-from typing import Optional
+from typing import Final, Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -17,9 +16,6 @@ class AssignableUserImageURI:
     
 
 
-class AssignableUserType(str, Enum):
-    USER = 'user'
-
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
@@ -27,7 +23,7 @@ class AssignableUserType(str, Enum):
 class AssignableUser:
     display_name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('display_name') }})
     org_id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('org_id') }})
-    type: AssignableUserType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    TYPE: Final[str] = dataclasses.field(default='user', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     image_uri: Optional[AssignableUserImageURI] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('image_uri'), 'exclude': lambda f: f is None }})
     user_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('user_id'), 'exclude': lambda f: f is None }})